synth: the public API speaks in bytes
Every MIDI event value is a byte on the wire, and the public entry
points now say so: channel, command, key, velocity and the data pair
all take u8, so the expected ranges are in the signatures rather than
in the source (upstream rustysynth issue #2 -- where note_on with a
velocity past i8::MAX went silent; ours has clamped for some while,
but the types still invited the mistake). The channel-taking accessor
family follows, one shape across the surface.

The engine still thinks in i32 past the door: each public method
widens on its first line, or hands straight to a private i32 worker
where the CC dispatch also calls in, so behaviour is bit-identical.
The engine's own call sites, whose values were bytes all along, drop
their up-casts.

+++ tests/modulator_behaviour.rs
@@ -11,7 +11,7 @@ use std::sync::Arc;
 
 use coppersynth::synth::{SoundFont, Synthesizer, SynthesizerSettings};
 
-fn render_note(font: &Arc<SoundFont>, velocity: i32) -> (Vec<f32>, Vec<f32>) {
+fn render_note(font: &Arc<SoundFont>, velocity: u8) -> (Vec<f32>, Vec<f32>) {
     let settings = SynthesizerSettings::new(44100);
     let mut synth = Synthesizer::new(font, &settings).unwrap();
     // Program 39: Synth Bass 2, the fixture patch. A low E, half a second.
